Move code-tour.ts to /code-tour
